[Refactor] 변수명 변경
