+++ app/src/main/java/org/sopt/at/home/component/HomeLazyRow.kt
@@ -20,12 +20,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import org.sopt.at.home.contents.Content
-import org.sopt.at.home.contents.Top20Contents
+import org.sopt.at.home.model.Content
+import org.sopt.at.home.model.Top20Contents
 
 @Composable
 fun HomeLazyRow(
